Actually request icons for AvatarList.astro.
Currently the images are hotlinked, so it might be the case where it makes more sense to pull in the images as assets programmatically and then reference them locally, but that's for once Astro has images figured out a bit more.

I fixed the editHref to actually point to this repo

There needs to be a Github Personal Access Token with `public_repo` permissions set to `SNOWPACK_PUBLIC_GITHUB_TOKEN` to remove Github rate limiting.