Make sure option details get wrapped in `<p>`
Turns out Markdown (remark? remark in Astro??) was deciding some lines were _not_ a paragraph if they contained an HTML tag — whether that were a `<br>` tag or the new `<Since />` component. This explicitly wraps the block containing the component in a `<p>` tag and also uses the more explicit `<br>` over two trailing spaces for clarity.

(An empty line is needed after the opening `<p>` to have the following text still parse as Markdown.)