A few stylistic changes
I hope I didn't change the meaning here, only meant to change to be more consistent with the voice we've been using throughout. If so, hit me up in the comments and we'll haggle over wording again!

But, this looks great! I know what the debugging component does now! I tested your example in my own repo and it works! So, time to get other eyes on this.