Refactor fallback content generation
Relies only on the existence of `.md` files to decide how to generate fallback pages rather than relying on the navigation contents as previously.

This is possible now that slugs are consistent across locales and helps for cases where a nav link isn’t for an actual page (“Installation” currently hits a redirect route for example).