Improve handling of RTL languages
Stores each language’s writing direction in the languages file to avoid needing to specify it in Markdown frontmatter as Arabic has done up until now.

This also allows us to layout RTL-language pages correctly even when they contain LTR fallback content. After #365 the site navigation was on the right on the Arabic “Getting Started” page, but on the left for fallback pages. Making this a per-language and not a per-page setting makes this consistent.

This commit also updates the `add-language` script to collect writing direction.