Use a component to render `import.meta.env` (#412)
This fixes the need to write `import[dot]meta[dot]env` by adding an `<ImportMetaEnv />` component to render the correct string for us.

This works but inline components seem to trip up Markdown parsing slightly so some paragraphs need wrapping in explicit `<p>` tags. (Each opening `<p>` is followed by an empty new line as otherwise the _rest_ of the paragraph won’t be parsed…)

It’s clunky on the authoring side, but it does mean we can provide much more readable docs.

H/T to @Princesseuh for the component idea.