Use language-specific link from docs logo
The docs logo at the top-left currently always links to `https://docs.astro.build/` which gets redirected to the English version of the docs. This change means that we link to the current language, e.g. `https://docs.astro.build/fr/` to keep users on their current language if they use that link.