Moves the Routing page's sidebar entry into Learn > Basics
This only moves where the Routing page shows up in the Left Sidebar.

We *do* want to move it into Learn.

Proposal: move it into "Basics" section?
 - NOT  "Features" because it's not an extra thing; it's fundamental
- maybe? "Core Concepts" because it's **so** fundamental?

Extra Context: We have a hidden placeholder for a (not-yet-written) "Why Astro?" page that will also go in Core Concepts. Would we consider Routing SO CORE as to belong with these items? Or just a "basic" item?