♻️ reword the dynamic JSX section
I found some of the headings in this section to be oddly worded, "dynamic" is used a lot here but it's also a bit misleading (for example I thought "Dynamic Attributes" at first meant actual dynamic HTML attributes like `<marklar dynamic-attribute-* />`. People coming from React or have used JSX know it's dynamic in nature so I think we can simplify the headings here a bit.

I also made minor changes to the sentence structure and wording of the descriptions for each heading.