♻️ clarify that a string or number can be used (#786)
* ♻️ clarify that a string or number can be used

It's a bit confusing that `id: 3` is a number whereas the other two are strings. When I first read this I thought it was a bug until I read above where it mentioned that strings and numbers can be used as parameters. I think it'd be nice to reflect that in the code block as well just to make it concrete.

* ♻️ move comment outside of curly braces

Co-authored-by: Sarah Rainsberger <sarah@rainsberger.ca>

* Update src/pages/en/reference/api-reference.md

Co-authored-by: Sarah Rainsberger <sarah@rainsberger.ca>

* removed an extra space

Co-authored-by: Sarah Rainsberger <sarah@rainsberger.ca>