Make RTL a top-level setting
Until now each Arabic page included a `dir: rtl` entry in its frontmatter. This commit adds a `rtlLanguages` set to our languages file so we can set this once for each language.

This means we can preserve the overall RTL layout even for fallback content where the page content is LTR but a user would still expect sidebars etc. to be consistent with their RTL expectations.