minor: Fix heading levels on docs homepage
a11y guidelines state that heading levels should increment by one level at a time (e.g. `<h4>` must be preceded by an `<h3>`) We were using `<h4>` for the first two headings on the docs homepage, which violates that rule. I suspect this was to keep these headings out of the sidebar or for stylistic reasons (smaller font size).

I’ve switched back to `<h2>` to fix this a11y issue (it impacts Lighthouse scores amongst other things), and used a literal `<h2>` tag for the first mini section to avoid that showing up in the table of contents.