Don’t store `i18nReadyDate` or commit messages in metadata
If the issue is that we’re hitting size limits on the issue body, it might help not to include this extra metadata which can be helpful for manual debugging but is not actually used by the action to update the dashboard.

`lastChange` is also not used for functionality but leaving that in for now as the most helpful of the surplus metadata.