Included common gotcha on script and style import (#2510)
* Included note on script and style import 

It can be hard to track down the source of a `<script>` or  `<style>` tag if it's included from an unused import. Also, it's important to point it out, as it might be expected that "non-explicitly used" JS isn't in the final build.

* Corrected "import" to "use"

Style-script bleed does not occur if component is only imported and not used.
It does occur if component is imported and used but not rendered (or maybe better "appears in the output", since it could be physically in the output HTML but `display:none` ).

* reviewed wording

i merged suggestions! still not sure about evaluated but i'm glad i'm not alone in this uncertainty 😁 

co-authored-by: sarah11918 <sarah@rainsberger.ca>
co-authored-by: Jutanium <danjutan@gmail.com>

* "used" -> included in the template

could this be our final iteration? 😁

* included intro

co-authored-by: sarah11918 <sarah@rainsberger.ca>

* Update src/pages/en/guides/troubleshooting.mdx

Co-authored-by: Dan Jutan <danjutan@gmail.com>

* made link relative

---------

Co-authored-by: sarah11918 <sarah@rainsberger.ca>
Co-authored-by: Jutanium <danjutan@gmail.com>