refactor: parameterize 'docs' entry types
Create and export a type `DocsEntryType`, which is the union of
allowable values in the 'type' field for an entry in the 'docs'
collection.

Create and export a type `DocsEntry<T>` where `T` is a `DocsEntryType`
and can be used to replace the type aliases like `TutorialEntry` and
`RecipeEntry`. Provides better LSP hover info by showing the types of
the 'data' field.

Create and export a function `createIsDocsEntry` which accepts a
parameter of type `DocsEntryType` and generates type guard functions for
use in `Array.prototype.filter`.

Refactor exported type aliases to use the `DocsEntry<T>` type, and
refactor exported type guard functions to call `createIsDocsEntry` to
define the type guard functions.