Update server-side-rendering.mdx
This commit makes the code example for adapters run when copy-pasted.

The current code example won't work as-is and throws a “Setting the 'mode' option is required.” error without any suggestions for resolution. Readers then have to realize that they need to go to docs for the nodejs adapter and figure out what `mode` is.