fix(markdown-content.mdx): fix incorrect remote markdown example
The original code may initially appear to work as `npm run dev` will re-fetch the content every time the page is refreshed, but running `npm run build` will reveal that in the actual behaviour, the content is only fetched at build time and is not in fact dynamically fetched during runtime.

The new code fixes this by moving the JavaScript outside of the component script and instead placing it in a `<script>` tag.

fix #7629