Update astro-db.mdx
add some missing details to make the tutorial more fluid.  
if you follow the tutorial, at this point, your comment table has author and content columns.  
The seed script is going to fail because it's missing data in the author table and the Comment table does not have body column