Add config file types to Project Structure page
We're working on ways to reorganize, or link to, the content in the
"Configuring Astro" guide.

As recommended in
https://github.com/withastro/docs/issues/8918#issuecomment-2255579753,
this PR will add the info on supported config file types on the
"Configuring Astro" guide to the "Project Structure" basics page. This
will give new users a quick reference when they're learning how to set
up their projects.