Fix grammatical error in Contributor Manual ("is has" → "has")
The sentence "This document is has some basic information to get you started..." contains a grammatical error due to the incorrect combination of "is" and "has" in this context. The word "is" is unnecessary and disrupts the sentence's structure.

The correct phrase should simply be "has", as it properly indicates possession of the information.