Update decap-cms.mdx
The admin dashboard won't work if the Decap CMS HTML file is created within the public folder. Astro won't generate the route. Instead it should be created only in `src/pages` directory. Using `.html` extension will allow the use of the `window.CMS` object, which is useful for customizing the admin page.