Update scoped styles documentation in guides
Previous docs state that scoped styles don't increase specificity which is inaccurate. Scoped styles will increase class specificity by 1.

The imported style in the second example will have a specificity of 0-0-2. This will not override the scoped style of 0-1-1. The example is updated to use an ID selector which will override the scoped style with its specificity of 1-0-0.