DeployHQ deployment guide FR version
Hi, this is my first contribution to translating docs !

I was wondering about translating the DeployHQ sections referenced in the deployment steps, for instance "Build pipeline", but I didn't because I figured DeployHQ doesn't seem to have a French version (I may be mistaken). So I could use a review for that, too. If needed, I'll translate.

Thanks !